+++ spec/mailers/clearance_mailer_spec.rb
@@ -55,4 +55,37 @@ describe ClearanceMailer do
       text: I18n.t("clearance_mailer.change_password.link_text")
     )
   end
+
+  context "when using a custom model" do
+    it "contains a link for a custom model" do
+      define_people_routes
+      Person = Class.new(User)
+      person = Person.new(email: "person@example.com", password: "password")
+
+      person.forgot_password!
+      host = ActionMailer::Base.default_url_options[:host]
+      link = "http://#{host}/people/#{person.id}/password/edit" \
+        "?token=#{person.confirmation_token}"
+
+      email = ClearanceMailer.change_password(person)
+
+      expect(email.text_part.body).to include(link)
+      expect(email.html_part.body).to include(link)
+
+      Object.send(:remove_const, :Person)
+      Rails.application.reload_routes!
+    end
+
+    def define_people_routes
+      Rails.application.routes.draw do
+        resources :people, controller: "clearance/users", only: :create do
+          resource(
+            :password,
+            controller: "clearance/passwords",
+            only: %i[edit update],
+          )
+        end
+      end
+    end
+  end
 end